+++ activity/brightpay/nextjs/changelog.md
@@ -5214,3 +5214,14 @@
 
 ---
 
+# Brightpay Development - 2025-04-12
+
+**Commit #328**
+
+- Repository: brightpay/brightclinic
+- Type: healthcare-platform
+- Language: nextjs
+- Month: 4/2025
+
+---
+